workflow/docker/requirements.txt
Flask==3.0.3
cryptography
Minio==7.2.5
python-dotenv==1.0.1
